move mnemonic conversion to a lib
