+++ src/lib/mod.rs
@@ -1,6 +1,7 @@
 //! All the common functionality.
 
 use anyhow::anyhow;
+use bip39::Mnemonic;
 use candid::{
     parser::typing::{check_prog, TypeEnv},
     types::Function,
@@ -13,8 +14,13 @@ use ic_agent::{
 use ic_base_types::PrincipalId;
 use ic_nns_constants::{GENESIS_TOKEN_CANISTER_ID, GOVERNANCE_CANISTER_ID, LEDGER_CANISTER_ID};
 use ic_types::Principal;
+use libsecp256k1::{PublicKey, SecretKey};
+use pem::{encode, Pem};
 use serde_cbor::Value;
-use std::convert::TryFrom;
+use simple_asn1::ASN1Block::{
+    BitString, Explicit, Integer, ObjectIdentifier, OctetString, Sequence,
+};
+use simple_asn1::{oid, to_der, ASN1Class, BigInt, BigUint};
 
 pub const IC_URL: &str = "https://ic0.app";
 
@@ -179,10 +185,50 @@ pub fn parse_query_response(
 }
 
 pub fn get_account_id(principal_id: Principal) -> AnyhowResult<ledger_canister::AccountIdentifier> {
+    use std::convert::TryFrom;
     let base_types_principal =
         PrincipalId::try_from(principal_id.as_slice()).map_err(|err| anyhow!(err))?;
     Ok(ledger_canister::AccountIdentifier::new(
         base_types_principal,
         None,
     ))
+}
+
+/// Converts menmonic to PEM format
+pub fn mnemonic_to_pem(mnemonic: &Mnemonic) -> String {
+    fn der_encode_secret_key(public_key: Vec<u8>, secret: Vec<u8>) -> Vec<u8> {
+        let secp256k1_id = ObjectIdentifier(0, oid!(1, 3, 132, 0, 10));
+        let data = Sequence(
+            0,
+            vec![
+                Integer(0, BigInt::from(1)),
+                OctetString(32, secret.to_vec()),
+                Explicit(
+                    ASN1Class::ContextSpecific,
+                    0,
+                    BigUint::from(0u32),
+                    Box::new(secp256k1_id),
+                ),
+                Explicit(
+                    ASN1Class::ContextSpecific,
+                    0,
+                    BigUint::from(1u32),
+                    Box::new(BitString(0, public_key.len() * 8, public_key)),
+                ),
+            ],
+        );
+        to_der(&data).expect("Cannot encode secret key.")
+    }
+
+    let seed = mnemonic.to_seed("");
+    let ext = tiny_hderive::bip32::ExtendedPrivKey::derive(&seed, "m/44'/223'/0'/0/0").unwrap();
+    let secret = ext.secret();
+    let secret_key = SecretKey::parse(&secret).unwrap();
+    let public_key = PublicKey::from_secret_key(&secret_key);
+    let der = der_encode_secret_key(public_key.serialize().to_vec(), secret.to_vec());
+    let pem = Pem {
+        tag: String::from("EC PRIVATE KEY"),
+        contents: der,
+    };
+    encode(&pem).replace("\r", "").replace("\n\n", "\n")
 }
